pyproject.toml
[project]
name = "podman-mcp"
version = "0.1.0"
description = "A Podman MCP server"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.0",
    "mcp>=1.0.0",
    "python-dotenv>=1.0.1",
    "podman>=4.9.0",
    "pyyaml>=6.0.1"
]

[[project.authors]]
name = "Alex Andru"
email = "alex007d@gmail.com"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
podman-mcp = "podman_mcp:main"